move contents of common into api
