+++ orderbook/src/api/get_order_by_uid.rs
@@ -1,4 +1,4 @@
-use crate::api::common::convert_get_orders_error_to_reply;
+use crate::api::convert_get_orders_error_to_reply;
 use crate::database::OrderFilter;
 use crate::orderbook::Orderbook;
 use anyhow::Result;
