+++ orderbook/src/api/get_orders.rs
@@ -1,5 +1,5 @@
 use super::H160Wrapper;
-use crate::api::common::convert_get_orders_error_to_reply;
+use crate::api::convert_get_orders_error_to_reply;
 use crate::database::OrderFilter;
 use crate::orderbook::Orderbook;
 use anyhow::Result;
